requirements.txt
mcdreforged>=2.12.0
msgraph-sdk>=1.4.0
qrcode>=7.4.2